add soft delete to conversation_message_participants
In new conversations, the list of people able to see a
particular message will no longer be equiavalent to the
conversation_participants, so we need to preserve who was able
to see messages so we can display this information in the
interface.

fixes CNVS-7129

test plan:
 - regression test old conversations; there should be no
   behavioral changes there. Especially look at:
   - deleting messages and conversations
   - submission comments (deleting these messages and then
     adding another comment)
   - the internal "delete_for_all" api call (support uses this
     to remove naughty conversations)
 - when deleting messages/conversations, verify with the
   console that the conversation_message_participants are still
   present with a workflow_state that is "deleted".

Reviewed-on: https://gerrit.instructure.com/22853
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Cam Theriault <cam@instructure.com>
Product-Review: Jon Willesen <jonw@instructure.com>

+++ app/models/conversation_participant.rb
@@ -200,7 +200,7 @@ class ConversationParticipant < ActiveRecord::Base
     }.with_indifferent_access
   end
 
-  def messages
+  def all_messages
     self.conversation.shard.activate do
       if self.conversation.shard == self.shard
         # use a slightly more forgiving backcompat query (since the migration may not have
@@ -218,6 +218,10 @@ class ConversationParticipant < ActiveRecord::Base
           order("created_at DESC, id DESC")
       end
     end
+  end
+
+  def messages
+    all_messages.where("(workflow_state <> ? OR workflow_state IS NULL)", 'deleted')
   end
 
   def participants(options = {})
@@ -263,18 +267,60 @@ class ConversationParticipant < ActiveRecord::Base
     conversation.add_message(user, body_or_obj, options.merge(:generated => false))
   end
 
+  # Public: soft deletes the message participants for this conversation
+  # participant for the specified messages. May pass :all to soft delete all
+  # message participants.
+  #
+  # to_delete - the list of messages to the delete
+  #
+  # Returns nothing.
   def remove_messages(*to_delete)
+    remove_or_delete_messages(:remove, *to_delete)
+  end
+
+  # Public: hard deletes the message participants for this conversation
+  # participant for the specified messages. May pass :all to hard delete all
+  # message participants.
+  #
+  # to_delete - the list of messages to the delete
+  #
+  # Returns nothing.
+  def delete_messages(*to_delete)
+    remove_or_delete_messages(:delete, *to_delete)
+  end
+
+  # Internal: soft or hard delete message participants, based on the indicated
+  # operation. Used by remove_messages and delete_messages methods.
+  #
+  # operation - The operation to perform.
+  #             :remove - Only set the workflow state on the message
+  #             participants.
+  #             :delete to delete the message participants from the database.
+  # to_delete - The list of conversation_messages to operate on. This function
+  #             only affects the conversation_message_participants for this
+  #             participant.
+  #
+  # Returns nothing.
+  def remove_or_delete_messages(operation, *to_delete)
     self.conversation.shard.activate do
       scope = ConversationMessageParticipant.joins(:conversation_message).
           where(:conversation_messages => { :conversation_id => self.conversation_id },
                           :user_id => self.user_id)
       if to_delete == [:all]
-        scope.delete_all
+        if operation == :delete
+          scope.delete_all
+        else
+          scope.update_all(:workflow_state => 'deleted')
+        end
       else
-        scope.where(:conversation_message_id => to_delete).delete_all
+        if operation == :delete
+          scope.where(:conversation_message_id => to_delete).delete_all
+        else
+          scope.where(:conversation_message_id => to_delete).update_all(:workflow_state => 'deleted')
+        end
         # if the only messages left are generated ones, e.g. "added
         # bob to the conversation", delete those too
-        return remove_messages(:all) unless messages.where(:generated => false).exists?
+        return remove_or_delete_messages(operation, :all) unless messages.where(:generated => false).exists?
       end
     end
     unless @destroyed
@@ -513,7 +559,7 @@ class ConversationParticipant < ActiveRecord::Base
 
   def destroy_conversation_message_participants
     @destroyed = true
-    remove_messages(:all) if self.conversation_id
+    delete_messages(:all) if self.conversation_id
   end
 
   def update_unread_count(direction=:up, user_id=self.user_id)
